requirement.txt
flask-cors-3.0.10
pandas
numpy==1.17.3
Flask==1.1.2
xlrd==2.0.1
openpyxl==3.0.5